Move the NTLM flags to a central location
